+++ internal/store/comments.go
@@ -18,7 +18,7 @@ type CommentStore struct {
 	db *sql.DB
 }
 
-func (s *CommentStore) GetByPostID(ctx context.Context, postID int) ([]Comment, error) {
+func (s *CommentStore) GetByPostID(ctx context.Context, postID int64) ([]Comment, error) {
 	query := `SELECT c.id, c.post_id, c.user_id, c.content, c.created_at, users.username, users.id FROM comments c
 	JOIN users on users.id = c.user_id
 	WHERE c.post_id = $1
